Add files via upload
This is what I have completed so far, for the final project in SDEV 140. The application I am creating is a hangman game. So far, the game is working I have two windows (a start window, and the main game window) I have a set list of words for the application to use. My next step is to get the application to actually draw the hangman as the user plays the game and guesses incorrectly.